Add fixes for GCC 12
