+++ NeoML/include/NeoML/TraditionalML/GraphGenerator.h
@@ -1,4 +1,4 @@
-/* Copyright © 2017-2020 ABBYY Production LLC
+/* Copyright © 2017-2023 ABBYY
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -65,8 +65,8 @@ private:
 	};
 	// The generating step
 	struct CStep {
-		int PathElementIndex; // the number of the path element from which the step starts
-		int ArcIndex; // the number of the arc in its starting node
+		int PathElementIndex = 0; // the number of the path element from which the step starts
+		int ArcIndex = 0; // the number of the arc in its starting node
 		Quality SumQuality; // the current path quality
 	};
 	typedef AscendingByMember<CStep, Quality, &CStep::SumQuality> StepQualityAscending;
